Address review: merge VLM into main vllm class, upgrade to v0.13
