+++ nemo_skills/pipeline/utils/server.py
@@ -24,7 +24,6 @@ LOG = logging.getLogger(get_logger_name(__file__))
 class SupportedServersSelfHosted(str, Enum):
     trtllm = "trtllm"
     vllm = "vllm"
-    vllm_vlm = "vllm_vlm"
     sglang = "sglang"
     megatron = "megatron"
     generic = "generic"
@@ -33,7 +32,6 @@ class SupportedServersSelfHosted(str, Enum):
 class SupportedServers(str, Enum):
     trtllm = "trtllm"
     vllm = "vllm"
-    vllm_vlm = "vllm_vlm"
     sglang = "sglang"
     megatron = "megatron"
     openai = "openai"
@@ -126,8 +124,8 @@ def get_server_command(
     num_tasks = num_gpus
 
     # check if the model path is mounted if not vllm, sglang, or trtllm;
-    # vllm, sglang, trtllm, and vllm_vlm can also pass model name as "model_path" so we need special processing
-    if server_type not in ["vllm", "vllm_vlm", "sglang", "trtllm", "generic"]:
+    # vllm, sglang, trtllm can also pass model name as "model_path" so we need special processing
+    if server_type not in ["vllm", "sglang", "trtllm", "generic"]:
         check_if_mounted(cluster_config, model_path)
 
     # the model path will be mounted, so generally it will start with /
@@ -160,8 +158,7 @@ def get_server_command(
             f"    --micro-batch-size 1 "  # that's a training argument, ignored here, but required to specify..
             f"    {server_args} "
         )
-    elif server_type in ["vllm", "vllm_vlm"]:
-        # vllm_vlm uses the same server as vllm; the difference is on the client side
+    elif server_type == "vllm":
         server_entrypoint = server_entrypoint or "-m nemo_skills.inference.server.serve_vllm"
         start_vllm_cmd = (
             f"python3 {server_entrypoint} "
